% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{continue.functions}
\alias{continue.functions}
\alias{C.fn.RBP}
\alias{C.fn.P}
\alias{C.fn.INST}
\alias{C.fn.RR}
\alias{C.fn.ERR}
\alias{C.fn.SDCG}
\alias{C.fn.AP}
\title{Continue functions for common IR metrics.}
\usage{
C.fn.RBP(p)

C.fn.P(k)

C.fn.INST(T)

C.fn.RR()

C.fn.ERR()

C.fn.SDCG(k)

C.fn.AP()
}
\arguments{
\item{p}{"Persistence" parameter for RBP.}

\item{k}{Cutoff or depth, beyond which the metric is not evaluated.}

\item{T}{"Target" parameter for INST.}
}
\description{
These are helper functions which build the \code{C.fn} parameter for
\code{\link{weighted.precision}}.
}
\details{
\code{weighted.precision} accepts not a \code{C} vector, but a function for
making one. This is because, for adaptive metrics, \code{C} will depend on
the gain (relevance) at each point down the ranked list and this relationship
can be arbitrarily complicated.

\code{weighted.precision} calls the \code{C} function with two parameters:
\code{R} is a vector of relevance, or gain. \code{i} is the current depth in
the rank list, i.e. the length of R. \code{C} should be vectorised over both
arguments.

The functions \code{C.fn.RBP} and \code{C.n.P} are parameterised in the same
way as the RBP and P metrics, by persistence \code{p} or cutoff \code{k}. The
function \code{C.fn.INST} is parameterised by \code{T}, the user's target
amount of relevance or gain.
}
