% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{IFT}
\alias{IFT}
\title{IR effectiveness metric derived from information foraging theory.}
\usage{
IFT(doc.gain, doc.cost, A, target, rationality, b1 = NA, intercept1 = NA,
  b2 = NA, intercept2 = NA)
}
\arguments{
\item{doc.gain}{A vector of gains from returned documents, typically (but not
always) each in the range 0..1.}

\item{doc.cost}{A vector of costs to examine SERP elements, typically (but not
always) each in the range 0..1.}

\item{A}{The allowed (minumum) rate of gain.}

\item{target}{The target (maximum) amount of gain. Called $T$ in the description
by Azzopardi et al.}

\item{rationality}{Controls the abruptness with which a searcher changes
strategy. Called $R1$ and $R2$ in Azzopardi et al., but here we use the same
value for both continue functions.}

\item{b1}{Base rate: rate of stopping when the searcher has enough gain.}

\item{intercept1}{Base rate: rate of stopping when the searcher has enough gain.}

\item{b2}{Base rate: rate of stopping when the searcher is getting gain at the allowable rate.}

\item{intercept2}{Base rate: rate of stopping when the searcher getting gain at the allowable rate.}
}
\value{
Compatible with other \link{ir.metrics} metrics, a list with the following elements:
\describe{
\item{metric}{The metric value.}
\item{C}{The computed \code{C} vector: \code{C[[i]]} is the probability of
the user Continuing to read past rank \emph{i}.}
\item{W}{The computed vector of weights: the importance placed on each rank
in the list.}
\item{L}{The computed vector of "last": the probability that each rank is
the last one the user examines.}
\item{gain}{The passed gains.}
\item{cum.metric}{The metric accumulated along the list.}
\item{i}{Ranks used (starting at 1).}
\item{C}{The computed \code{C} vector: \code{C[[i]]} is the probability of
the user Continuing to read past rank \emph{i}.}
}

IFT adds:\describe{
\item{C1}{The computed \code{C1} vector: \code{C1[[i]]} is the probability of
the user Continuing to read past rank \emph{i}, based on accumulated gain.}
\item{C2}{The computed \code{C2} vector: \code{C2[[i]]} is the probability of
the user Continuing to read past rank \emph{i}, based on rate of gain.}
\item{cost}{The passed costs.}
\item{rate}{The rate of gain at each rank.}
}
}
\description{
Note that unlike the other metrics in this package, \code{IFT} takes as
input both \emph{gain} and \emph{cost} vectors.
}
\details{
For details, see Azzopardi et al., "Measuring the Utility of Search Engine Result
Pages", Proc. SIGIR 2018.

\code{b1} and \code{b2} can be specified either directly (via the parameters of
the same name), or via intercepts: where the intercept is the probability of stopping
(C1) or continuing (C2) when the searcher is exactly at their target gain or
allowable rate. If neither \code{b2} nor \code{intercept2} is given, values are copied
from \code{b1}.
}
\seealso{
\link{ir.metrics}
}
