% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ir.metrics}
\alias{ir.metrics}
\alias{RBP}
\alias{P}
\alias{RR}
\alias{ERR}
\alias{SDCG}
\alias{AP}
\alias{INST}
\title{IR effectiveness metrics in the "weighted precision" family.}
\usage{
RBP(doc.gain, p)

P(doc.gain, k)

RR(doc.gain)

ERR(doc.gain)

SDCG(doc.gain, k)

AP(doc.gain)

INST(doc.gain, T)
}
\arguments{
\item{doc.gain}{A vector of gains from returned documents, typically (but not
always) each in the range 0..1.}

\item{p}{For \code{RBP}, the persistence parameter.}

\item{k}{For \code{P}, the cutoff rank.}

\item{T}{For \code{INST}, the user's target (number of relevant documents, or
total gain).}
}
\value{
All these metrics return a list with the following elements:
\describe{
\item{metric}{The metric value.}
\item{C}{The computed \code{C} vector: \code{C[[i]]} is the probability of
the user Continuing to read past rank \emph{i}.}
\item{W}{The computed vector of weights: the important placed on each rank
in the list.}
\item{gain}{The passed gains.}
\item{cum.metric}{The metric accumulated along the list.}
\item{i}{Ranks used (starting at 1).}
}

For RBP and INST, there will be a further element \code{residual} which
is the total gain possible from unjudged documents: i.e., \code{metric+residual}
is an upper bound on effectiveness if all unjudged documents had gain=1.

For details of this approach, see Moffat et al., "Users Versus Models: What
Observation Tells Us About Effectiveness Metrics", in Proc. Conf. Information
and Knowledge Management (CIKM) 2013.
}
\description{
These common IR metrics take a vector of "gains", which represents the
quality (or relevance) of each item in a search engine's list of retrieved
documents, and return a measure of the effectiveness of that list.
}
\details{
\code{RBP} is rank-biased precision, with persistence parameter p.

\code{P} is precision at depth k.

\code{RR} is reciprocal rank.

\code{ERR} is expected (graded) reciprocal rank. (Chapelle et al., Proc
CIKM'09: https://doi.org/10.1145/1645953.1646033.)

\code{SDCG} is (scaled) discounted cumulative gain at depth k, an
analogue of NDCG. (Moffat et al., TOIS 35(3):
https://doi.org/10.1145/3052768.)

\code{AP} is average precision.

\code{INST} is an adaptive and goal-sensitive metric which has different
behaviour depending whether the user is looking for few or many documents,
and depending on the quality of the list. (Moffat et al., TOIS 35(3):
https://doi.org/10.1145/3052768.)
}
\examples{
RBP(c(1, 1, 1, 0, 0), p=0.6)
# 0.8501041
P(c(1, 1, 1, 0, 0), k=1)
# 1
P(c(1, 1, 1, 0, 0), k=5)
# 0.6
RR(c(0, 0, 0, 1))
# 0.25
ERR(c(0, 1, 0))
# 0.5
ERR(c(0, .5, 0))
# 0.2
SDCG(c(1, 0, 1, 0, 0), k=1)
# 1
SDCG(c(1, 0, 1, 0, 0), k=5)
# .5087403
AP(c(1, 0, 1, 0, 0))
# 0.8333333
INST(c(1, 1, 1, 0, 0), T=1)
# 0.9830949
}
\seealso{
\code{\link{IFT}}
}
