% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{weighted.precision}
\alias{weighted.precision}
\title{Generic weighted-precision metric for IR effectiveness.}
\usage{
weighted.precision(C.fn, doc.gain, minN = 1)
}
\arguments{
\item{C.fn}{A function which, given the arguments \code{R} and
vector \code{i}, returns the \code{C} parameter defining the metric. See
Moffat et al. for details (or see examples in this package such as
\code{\link{C.fn.RBP}}).}

\item{doc.gain}{A vector of gains from documents, typically each in the range
0..1. Also sometimes known as \code{r} or the "relevance" vector.}

\item{minN}{The minimum length of list (number of documents) for which the
metric makes sense. \code{doc.gain} vectors shorter than this will be padded
with 0.}
}
\description{
\code{weighted.precision} takes a metric as defined by a \code{C} (continue)
vector, a vector \code{doc.gain} of gains from documents (typically 0..1 for
each), and returns the value of the metric.
}
\details{
This function can be used to calculate any number of metrics in the "weighted
precision" family, which includes precision (P@k), AP, RBP, MRR, and INST. A
metric is entirely defined by a vector \code{C}, where each element
\code{C[i]} gives the probability of a user who is looking at rank \code{i}
continuing on to rank \code{i+1}.

For details of this approach, see Moffat et al., "Users Versus Models: What
Observation Tells Us About Effectiveness Metrics", in Proc. Conf. Information
and Knowledge Management (CIKM) 2013.

You may find the derived functions \code{\link{P}}, \code{\link{RBP}}, and
\code{\link{INST}} easier to use.
}
